pyproject.toml
# pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mosaicmap"
version = "0.1.0"
authors = [
    { name="Eric Langenberg", email="eric-langenberg@users.noreply.github.com" },
]
description = "A Python library for creating treemaps with images"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Pillow>=9.0.0",
    "numpy>=1.20.0",
    "squarify>=0.4.3",
    "pandas>=1.3.0",
]

[project.urls]
"Homepage" = "https://github.com/eric-langenberg/mosaicmap"
"Bug Tracker" = "https://github.com/eric-langenberg/mosaicmap/issues"